fix an type error

+++ src/dc_chat.rs
@@ -784,7 +784,7 @@ unsafe fn get_parent_mime_headers(
                 "SELECT rfc724_mid, mime_in_reply_to, mime_references \
                  FROM msgs WHERE chat_id=? AND timestamp=(SELECT max(timestamp) \
                  FROM msgs WHERE chat_id=? AND from_id!=?);",
-                params![(*chat).id as i32, (*chat).id as i32, DC_CONTACT_ID_SELF],
+                params![(*chat).id as i32, (*chat).id as i32, DC_CONTACT_ID_SELF as i32],
                 |row| {
                     *parent_rfc724_mid = row.get::<_, String>(0)?.strdup();
                     *parent_in_reply_to = row.get::<_, String>(1)?.strdup();
@@ -802,7 +802,7 @@ unsafe fn get_parent_mime_headers(
                     "SELECT rfc724_mid, mime_in_reply_to, mime_references \
                      FROM msgs WHERE chat_id=? AND timestamp=(SELECT min(timestamp) \
                      FROM msgs WHERE chat_id=? AND from_id==?);",
-                    params![(*chat).id as i32, (*chat).id as i32, DC_CONTACT_ID_SELF],
+                    params![(*chat).id as i32, (*chat).id as i32, DC_CONTACT_ID_SELF as i32],
                     |row| {
                         *parent_rfc724_mid = row.get::<_, String>(0)?.strdup();
                         *parent_in_reply_to = row.get::<_, String>(1)?.strdup();
